datapath: move CompilationLock to datapath/loader/types.CompilationLock
